Split and align mHC Pallas kernel modules
