+++ src/maxtext/configs/base.yml
@@ -1306,7 +1306,7 @@ enable_mhc_lite: False
 # enable_mhc_lite=True, mhc_expansion_rate=4, bfloat16 activations, tensor
 # parallelism of 1, and a per-device token count divisible by
 # mhc_pallas_block_size.
-use_mhc_pallas_kernel: False
+enable_mhc_pallas_kernel: False
 # Token-axis (T = batch * sequence) block size. Must be a multiple of 8.
 mhc_pallas_block_size: 64
